:recycle: Remove unused variable, functions and components
