+++ libs/wcm/constants/permissions.js
@@ -2,8 +2,3 @@ export const SIGNING_METHODS = {
   SIGN_TRANSACTION: { key: 'sign_transaction', title: 'Signature request' },
   SIGN_MESSAGE: { key: 'sign_message', title: 'Sign message' },
 };
-
-export const EVENTS = {
-  TRANSACTION_MADE: 'transaction_made',
-  TOKEN_RECEIVED: 'token_received',
-};
