+++ libs/wcm/hooks/useSession.js
@@ -25,8 +25,7 @@ export const useSession = () => {
 
     await Promise.all(
       client.session.keys.map(async (key, index) => {
-        const session = client.session.get(key);
-        loadedSessions[index] = session;
+        loadedSessions[index] = client.session.get(key);
       })
     );
 
@@ -122,7 +121,9 @@ export const useSession = () => {
 
   useEffect(() => {
     if (client?.session && !hasLoaded) {
-      loadSessions();
+      (async () => {
+        await loadSessions();
+      })();
     }
   }, [client, sessions]);
