--- packages/views/screens/managers/initialization/initialization.css
@@ -1,39 +1,0 @@
-@import '../../../setup/react/app/mixins.css';
-
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  max-width: 550px;
-  margin: 0 auto;
-  height: 100%;
-}
-
-.headerIcon {
-  height: 87px;
-}
-
-.header {
-  @mixin headingLarge;
-
-  margin-bottom: 15px;
-  color: var(--color-maastricht-blue);
-}
-
-.text {
-  @mixin contentLargest;
-
-  text-align: center;
-  margin-bottom: 30px;
-  color: var(--color-slate-gray);
-}
-
-.link {
-  color: var(--color-link-active);
-  cursor: pointer;
-}
-
-.button {
-  width: 430px;
-}
