+++ setup/jest.config.js
@@ -134,7 +134,6 @@ module.exports = {
     'src/utils/analytics.js',
     'src/modules/bookmark/utils.js',
     'src/theme/Theme.js',
-    'src/utils/platform.js',
     'src/modules/wallet/utils/hwManager.js',
     'src/modules/wallet/utils/account.js',
     'src/utils/dateTime.js',
@@ -160,7 +159,6 @@ module.exports = {
     'src/modules/wallet/utils/api/index.js',
     'src/modules/transaction/utils/hwManager/index.js',
     'src/modules/transaction/utils/transactionDetailsHelper.js',
-    'src/modules/transaction/utils/removeDuplicateTransactions.js',
     'src/modules/block/map/blocksFiltersMap.js',
     'src/modules/block/utils/blockSubscribe.js',
     'src/modules/wallet/components/signMultisigView/helpers.js',
