+++ src/const/queries.js
@@ -1,7 +1,4 @@
-export const APPLICATION = 'BLOCKCHAIN_APPLICATION';
-export const LATEST_BLOCKS = 'GET_LATEST_BLOCKS';
 export const BLOCKS = 'GET_BLOCKS';
-export const BLOCKS_ASSETS = 'GET_BLOCKS_ASSETS';
 export const NETWORK_STATUS = 'GET_NETWORK_STATUS';
 export const LEGACY = 'GET_LEGACY';
 export const TOKENS_BALANCE = 'GET_TOKENS_BALANCE';
@@ -10,7 +7,6 @@ export const TOKENS_TOP_LSK_BALANCE = 'GET_TOKENS_TOP_LSK_BALANCE';
 export const TOKENS_ACCOUNT_EXISTS = 'GET_TOKENS_ACCOUNT_EXISTS';
 export const PEERS = 'GET_PEERS';
 export const VALIDATORS = 'GET_VALIDATORS';
-export const VALIDATOR = 'GET_VALIDATOR';
 export const STAKES_SENT = 'GET_STAKES_SENT';
 export const STAKES_RECEIVED = 'GET_STAKES_RECEIVED';
 export const UNLOCKS = 'GET_UNLOCKS';
@@ -24,7 +20,6 @@ export const TRANSACTION_STATISTICS = 'GET_TRANSACTION_STATISTICS';
 export const NETWORK_STATISTICS = 'GET_NETWORK_STATISTICS';
 export const EVENTS = 'GET_TRANSACTION_EVENTS';
 export const BLOCKCHAIN_APPS = 'GET_ALL_BLOCKCHAIN_APPS';
-export const CCM = 'GET_CCM';
 export const COMMAND_PARAMETERS_SCHEMAS = 'GET_COMMAND_PARAMETERS_SCHEMAS';
 export const FEES = 'GET_FEES';
 export const BLOCKCHAIN_APPS_STATISTICS = 'GET_BLOCKCHAIN_APPS_STATISTICS';
@@ -37,5 +32,3 @@ export const REWARD_DEFAULT = 'GET_REWARD_DEFAULT';
 export const REWARD_INFLATION = 'GET_REWARD_INFLATION';
 export const REWARD_CONSTANTS = 'GET_REWARD_CONSTANTS';
 export const INVOKE = 'POST_PROXY_INVOKE';
-export const INITIALIZED_ACCOUNTS = 'GET_INITIALIZED_ACCOUNTS';
-export const MAIN_CHAIN_APPLICATIONS_BY_NETWORKS = 'GET_MAIN_CHAIN_APPLICATIONS_BY_NETWORKS';
