+++ src/modules/account/const/index.js
@@ -1,1 +1,3 @@
+export const defaultDerivationPath = "m/44'/134'/0'";
+
 export * from './accountMenu';
