--- src/modules/search/utils/index.js
@@ -1,2 +1,0 @@
-export * from './searchUtils';
-export { default as getTxConfig } from '@transaction/utils/transactionListUtils';
