+++ src/modules/transaction/configuration/transactions.js
@@ -2,7 +2,6 @@ const maxMessageLength = 64;
 const MIN_ACCOUNT_BALANCE = 5000000;
 const MIN_FEE_PER_BYTE = 1000;
 
-const DUMMY_RECIPIENT_ADDRESS = 'lskdxc4ta5j43jp9ro3f8zqbxta9fn6jwzjucw7yt';
 const DEFAULT_NUMBER_OF_SIGNATURES = 1;
 const DEFAULT_SIGNATURE_BYTE_SIZE = 64;
 
@@ -13,7 +12,6 @@ export {
   MIN_ACCOUNT_BALANCE,
   MIN_FEE_PER_BYTE,
   DEFAULT_NUMBER_OF_SIGNATURES,
-  DUMMY_RECIPIENT_ADDRESS,
   DEFAULT_SIGNATURE_BYTE_SIZE,
   STAKE_AMOUNT_STEP,
 };
