+++ src/modules/transaction/utils/index.js
@@ -2,6 +2,5 @@ export { default as normalizeStakesForTx } from './normalizeStakesForTx';
 export * from './moduleCommand';
 export * from './transaction';
 export * from './transactionDetailsHelper';
-export * from './removeDuplicateTransactions';
 export * from './showNotificationsForIncomingTransactions';
 export * from './encoding';
