--- src/modules/transaction/utils/removeDuplicateTransactions.js
@@ -1,5 +1,0 @@
-export const removeDuplicateTransactions = (pendingTransactions, confirmedTransactions) =>
-  [...pendingTransactions, ...confirmedTransactions].filter(
-    (transactionA, index, self) =>
-      index === self.findIndex((transactionB) => transactionB.id === transactionA.id)
-  );
