+++ src/modules/wallet/utils/hwManager.js
@@ -78,11 +78,4 @@ const signMessageByHW = async ({ account, message }) => {
   }
 };
 
-export const getDeviceType = (deviceModel = '') => {
-  if (/ledger/i.test(deviceModel)) {
-    return 'ledgerNano';
-  }
-  return '';
-};
-
 export { updateTransactionSignatures, signMessageByHW };
